start work on loading image when result found
