--- src/FX/ControllerTest.java
@@ -1,8 +1,0 @@
-package FX;
-
-class ControllerTest {
-
-    @org.junit.jupiter.api.AfterEach
-    void tearDown() {
-    }
-}
