+++ src/fx/ControllerTest.java
@@ -1,0 +1,8 @@
+package fx;
+
+class ControllerTest {
+
+    @org.junit.jupiter.api.AfterEach
+    void tearDown() {
+    }
+}
